fix: move some components to their own folders
