+++ src/components/DpColumnSelector/DpColumnSelector.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs'
-import DpColumnSelector from './DpColumnSelector'
+import DpColumnSelector from './'
 
 <Meta
     title="Components/ColumnSelector"
