+++ src/components/DpColumnSelector/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DpColumnSelector.vue'
+export * from './DpColumnSelector.vue'
