+++ src/components/DpDataTable/DpDataTable.stories.mdx
@@ -1,4 +1,4 @@
-import { Canvas, Meta } from '@storybook/blocks'
+import { Meta } from '@storybook/addon-docs'
 import DpDataTable from './'
 
 <Meta
