+++ src/components/DpDataTableExtended/DpDataTableExtended.stories.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/addon-docs'
-import DpDataTableExtended from './DpDataTableExtended'
+import DpDataTableExtended from './'
 
 <Meta
     title="Components/DataTableExtended"
