+++ src/components/DpDataTableExtended/index.ts
@@ -1,0 +1,2 @@
+export { default } from './DpDataTableExtended.vue'
+export * from './DpDataTableExtended.vue'
