+++ src/components/index.js
@@ -1,4 +1,4 @@
-import dataTableSearch from './DpDataTable/DataTableSearch'
+import dataTableSearch from './DpDataTableExtended/DataTableSearch'
 import DpAccordion from './DpAccordion'
 import DpAutocomplete from './DpAutocomplete'
 import DpButton from './DpButton'
@@ -8,10 +8,10 @@ import DpBadge from './DpBadge'
 import DpCard from './DpCard'
 import DpCheckbox from './DpCheckbox'
 import DpCheckboxGroup from './DpCheckboxGroup'
-import DpColumnSelector from './DpDataTable/DpColumnSelector'
+import DpColumnSelector from './DpColumnSelector'
 import DpContextualHelp from './DpContextualHelp'
 import DpDataTable from './DpDataTable'
-import DpDataTableExtended from './DpDataTable/DpDataTableExtended'
+import DpDataTableExtended from './DpDataTableExtended'
 import DpDatepicker from './DpDatepicker'
 import DpDateRangePicker from './DpDateRangePicker'
 import DpDatetimePicker from './DpDatetimePicker'
@@ -34,7 +34,7 @@ import DpRadio from './DpRadio'
 import DpResettableInput from './DpResettableInput'
 import DpSearchField from './DpSearchField'
 import DpSelect from './DpSelect'
-import DpSelectPageItemCount from './DpDataTable/DpSelectPageItemCount'
+import DpSelectPageItemCount from './DpDataTableExtended/DpSelectPageItemCount'
 import DpSkeletonBox from './DpSkeletonBox'
 import DpSlidebar from './DpSlidebar'
 import DpSlidingPagination from './DpSlidingPagination'
