+++ tests/DataTableSearch.spec.js
@@ -1,4 +1,4 @@
-import DataTableSearch from '~/components/DpDataTable/DataTableSearch'
+import DataTableSearch from '~/components/DpDataTableExtended/DataTableSearch'
 
 describe('DataTableSearch a utility function to perform full-text search in strings', () => {
   it('should escape all regex special chars', () => {
